html/semantics/forms/the-input-element/keypress-link-trigger.tentative.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>Do not trigger parent link when typing `Enter` in a single line text control</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
"use strict";

addEventListener("load", () => {
  promise_test(async () => {
    const form = document.querySelector("form");
    let submitted = false;
    form.addEventListener("submit", event => {
      submitted = true;
      event.preventDefault();
    });
    const textControl = document.querySelector("input[type=text]");
    textControl.focus();
    // typing `Enter` to submit the form
    await new test_driver.Actions()
      .keyDown("\uE007")
      .keyUp("\uE007")
      .send();
    assert_true(submitted, "Typing `Enter` should've caused submitting the form");
    assert_equals(
      document.querySelector("span:target"),
      null,
      "The parent link should not be performed by typing `Enter` to submit the form"
    );
  });
}, {once: true});
</script>
</head>
<body>
  <form action="#submit">
    <a href="#anchor">
      <input type="text">
      <input type="submit">
    </a>
  </form>
  <span id="anchor">targeted if the link is performed</span>
</body>
</html>
